+++ OgreMain/include/OgreSmallVector.h
@@ -374,7 +374,7 @@ namespace Ogre
             // Use memcpy for PODs iterated by pointers (which includes SmallVector
             // iterators): std::uninitialized_copy optimizes to memmove, but we can
             // use memcpy here.
-            memcpy( Dest, I, ( E - I ) * sizeof( T ) );
+            memcpy( Dest, I, static_cast<size_t>( E - I ) * sizeof( T ) );
         }
 
         /// grow - double the size of the allocated memory, guaranteeing space for at
